Integrate video (#28)
* termcam

* fmt

* gate stream video behind a feature flag

* clippy

* test all feature except uitest on ci

* update all deps

* add a boder around the framebuffer

* fmt

* improve compile time, keep the same verison of crossterm in termchat and tui

* Install with all features

* rename to startstream

* libxkbcommon is not needed anymore

* ci: build/package with --all-features

* ci: clippy --all-features, +fmt code

+++ src/message.rs
@@ -9,9 +9,9 @@ pub enum Chunk {
 
 #[derive(Serialize, Deserialize)]
 pub enum NetMessage {
-    HelloLan(String, u16),                    // user_name, server_port
-    HelloUser(String),                        // user_name
-    UserMessage(String),                      // content
-    UserData(String, Chunk),                  // file_name, chunk
-    Stream(Option<(Vec<u32>, usize, usize)>), // Option of (stream_data width, height ) None means stream has ended
+    HelloLan(String, u16),                   // user_name, server_port
+    HelloUser(String),                       // user_name
+    UserMessage(String),                     // content
+    UserData(String, Chunk),                 // file_name, chunk
+    Stream(Option<(Vec<u8>, usize, usize)>), // Option of (stream_data width, height ) None means stream has ended
 }
